Split registry into color-picker / gradient-picker / fill-picker (v1.1.0)
Color-only consumers can now install just `color-picker.json` (21 files)
instead of pulling all 53 files of the unified bundle. Gradient consumers
install `gradient-picker.json` (27 own files + color-picker via
registryDependencies). The full `fill-picker.json` (6 own files + the
other two) stays as the "give me everything" default and is what the
hero install snippet still points to.

Adds gradient-picker.tsx mirroring fill-picker.tsx's namespace pattern.
fill-picker.tsx now delegates to gradient-picker.tsx to keep exports DRY.
build-registry.ts carries the new per-item `version` field through to
the emitted JSON. Docs Installation section now shows all three entry
points with what-you-get descriptions.
