+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "amplo-fill-picker",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "private": true,
   "type": "module",
   "packageManager": "pnpm@9.15.9",
