+++ registry/new-york/color-picker/fill-picker.tsx
@@ -1,98 +1,23 @@
 "use client";
 
-// Re-export everything ColorPicker exposes so a single import point covers
-// the whole fill-picker public surface.
-export * from "./color-picker";
-export { ColorPicker } from "./color-picker";
+// Re-export the gradient-picker public surface (which itself re-exports
+// color-picker), so a single import point covers the whole fill-picker
+// public surface: color + gradient + the fill switcher below.
+export * from "./gradient-picker";
+export { ColorPicker, GradientPicker } from "./gradient-picker";
 
-// FillPicker shell
 import { Root as FillRoot } from "./parts/fill/root";
 import { Tabs as FillTabs, Tab as FillTab } from "./parts/fill/tabs";
 import { Pane as FillPane } from "./parts/fill/pane";
 
-// GradientPicker parts
-import { Root as GradientRoot } from "./parts/gradient/root";
-import { Bar } from "./parts/gradient/bar";
-import { Area as GradientArea } from "./parts/gradient/area";
-import { Overlay as GradientOverlay } from "./parts/gradient/overlay";
-import { TypeSwitcher } from "./parts/gradient/type-switcher";
-import { ReverseStops } from "./parts/gradient/reverse-stops";
-import { RepeatingToggle } from "./parts/gradient/repeating-toggle";
-import { AnglePad } from "./parts/gradient/angle-pad";
-import { AngleInput } from "./parts/gradient/angle-input";
-import { PositionPad } from "./parts/gradient/position-pad";
-import { PositionInput } from "./parts/gradient/position-input";
-import { ShapeSwitcher } from "./parts/gradient/shape-switcher";
-import { RadiusInput } from "./parts/gradient/radius-input";
-import { EllipseRadiiInput } from "./parts/gradient/ellipse-radii-input";
-import { RadialSizeSelect } from "./parts/gradient/radial-size-select";
-import { StopList } from "./parts/gradient/stop-list";
-import { StopColor } from "./parts/gradient/stop-color";
-import { InterpSwitcher } from "./parts/gradient/interp-switcher";
-import { Presets, BUILTIN_GRADIENT_PRESETS } from "./parts/gradient/presets";
-import { CssInput as GradientCssInput } from "./parts/gradient/css-input";
-import { PositionGroup } from "./parts/gradient/position-group";
-import { AngleGroup } from "./parts/gradient/angle-group";
-
-export type {
-  Fill,
-  ColorFill,
-  GradientFill,
-  Gradient,
-  GradientType,
-  GradientInterp,
-  GradientStop,
-  LinearGradient,
-  RadialGradient,
-  RadialSizeKeyword,
-  ConicGradient,
-} from "./lib/gradient";
-export {
-  formatFill,
-  parseFill,
-  formatGradient,
-  parseGradient,
-  DEFAULT_LINEAR,
-  DEFAULT_RADIAL,
-  DEFAULT_CONIC,
-} from "./lib/gradient";
-export { BUILTIN_GRADIENT_PRESETS };
-export { useGradientPicker } from "./hooks/use-gradient-picker";
-export type {
-  UseGradientPickerProps,
-  GradientPickerState,
-} from "./hooks/use-gradient-picker";
+export type { Fill, ColorFill, GradientFill } from "./lib/gradient";
+export { formatFill, parseFill } from "./lib/gradient";
 export { useFillPicker } from "./hooks/use-fill-picker";
 export type {
   UseFillPickerProps,
   FillPickerState,
   FillMode,
 } from "./hooks/use-fill-picker";
-
-export const GradientPicker = {
-  Root: GradientRoot,
-  Bar,
-  Area: GradientArea,
-  Overlay: GradientOverlay,
-  TypeSwitcher,
-  ReverseStops,
-  RepeatingToggle,
-  AnglePad,
-  AngleInput,
-  PositionPad,
-  PositionInput,
-  ShapeSwitcher,
-  RadiusInput,
-  EllipseRadiiInput,
-  RadialSizeSelect,
-  StopList,
-  StopColor,
-  InterpSwitcher,
-  Presets,
-  CssInput: GradientCssInput,
-  PositionGroup,
-  AngleGroup,
-};
 
 export const FillPicker = {
   Root: FillRoot,
